refactor: 경로 패키지 분리
