+++ src/main/java/chess/domain/path/MovablePaths.java
@@ -1,5 +1,6 @@
-package chess.domain;
+package chess.domain.path;
 
+import chess.domain.Position;
 import java.util.List;
 
 public class MovablePaths {
