+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -2,11 +2,9 @@ package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.MovablePaths;
-import chess.domain.Path;
+import chess.domain.path.MovablePaths;
 import chess.domain.Position;
 import chess.domain.TeamColor;
-import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
